Update ultrasonic_sensor_fan.ino
How does it work? 
This module consists of two drums, one of which is an emitter that emits ultrasound and other is receiver which receives the reflected ultrasound from the object. The emitter drum emits ultrasound when we trigger the module using the trigPin by sending a 10 microseconds high pulse. The module makes the echoPin high. Emitted ultrasound travels forward till it gets reflected by an object. When the reflected ultrasound is received by the receiver, echoPin is made low. This is how we get the time and multiple by the speed we get the distance.